Spec: Remove conditional macro def for %license
In the %files section, "COPYING" and "COPYING.LESSER" are marked with
the "%license" tag to facilitate special handling by rpm (e.g., the
"--licensefiles" option).

When this tag was added to "munge.spec" in commit 2b13519, support
was not as ubiquitous so a conditional macro definition was added
to allow the tag to function with older versions of rpm.  That macro
definition is no longer necessary since the current spec file targets
RHEL7+ and recent Fedora releases.

Remove the conditional macro definition for %license.

Reference:
- https://docs.fedoraproject.org/en-US/packaging-guidelines/LicensingGuidelines/
- http://rpm.org/ticket/116

Tested:
- AlmaLinux 9.0, 8.6
- CentOS Linux Stream 9, Stream 8, 8.5.2111, 7.9.2009
- Fedora 36, 35, 34