libmunge: Fix err msg suggestion for client ENOENT
In libmunge, _m_msg_client_connect() stat()s the munged listening
socket before attempting connect().  An error is returned to the
client if the stat() fails.  In commit f56acfe, this error message was
changed for the case of ENOENT to add a suggestion to start munged.
However, that change was not limited to just ENOENT.  Consequently,
EACCES (permission denied) now returns this suggestion to start munged
which is wrong.

Fix the error message to only suggest starting munged for the case
of ENOENT.

Tested:
- Debian GNU/Linux bookworm/sid