Sharness: Remove munge user in 1000-chaos-rpm.t
The "munge" user and group are created in the %pre section of the
specfile which is run when the munge RPM is installed.

Remove the munge user (which will also remove its group) at the end
of the test in order to ensure the user and group will be recreated
when the RPM is installed.

Tested:
- AlmaLinux 9.1, 8.7
- CentOS Linux Stream 9, Stream 8, 8.5.2111, 7.9.2009
- Fedora 37, 36, 35