Sharness: Set SHARNESS_TEST_OUTDIR to CWD for VPATH
If SHARNESS_TEST_OUTDIR is not defined, VPATH builds can clobber the
SHARNESS_TRASH_DIRECTORY making it difficult to diagnose failures in
the test suite.

The SHARNESS_TRASH_DIRECTORY is the HOME directory for a given
sharness test script, and it is recreated at the start of the script's
execution.  If the "root" variable is defined, it is prepended
to this directory pathname.  Then, if the trash directory is a
relative pathname (as it is in the default case), it is prepended
with SHARNESS_TEST_OUTDIR.  If SHARNESS_TEST_OUTDIR is not defined,
it defaults to SHARNESS_TEST_DIRECTORY, the directory containing the
sharness test script in the source tree.  This results in VPATH builds
clobbering the same trash directory for a given test script.

Prevent VPATH builds from clobbering the trash directory by setting
the current working directory as the default SHARNESS_TEST_OUTDIR.
This relocates the default trash directory from the source tree to
the build tree.

Tested:
- AlmaLinux 9.2, 8.7
- Arch Linux
- CentOS Linux Stream 9, Stream 8, 8.5.2111, 7.9.2009, 6.10
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Debian 12.0, 11.7, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 38, 37, 36
- FreeBSD 13.2, 13.1, 12.4
- NetBSD 9.3, 9.2, 9.1
- OpenBSD 7.3, 7.2, 7.1
- openSUSE 15.4, 15.3, 15.2, 15.1
- Ubuntu 23.04, 22.04.2, 20.04.6, 18.04.6, 16.04.7, 14.04.6, 12.04.5