Sharness: Disable %source_date_epoch_from_changelog for rpmbuild
"munge.spec.in" does not contain a %changelog since this information
is maintained in the commit log.  As of RPM 4.18.0 (2022-09-20),
rpmbuild will generate the following warning:

  %source_date_epoch_from_changelog is set, but %changelog has no
  entries to take a date from

As the name suggests, this RPM macro sets SOURCE_DATE_EPOCH from the
latest %changelog entry.

Instead of disabling this macro in "~/.rpmmacros", define this on
the rpmbuild command-line in "tests/1000-chaos-rpm.t".

Reference:
- https://rpm.org/wiki/Releases/4.18.0
- https://rpm-software-management.github.io/rpm/manual/buildprocess.html

Tested:
- Fedora 41, 40