Sharness: Remove echo options
Testing on macos-latest uncovered the following failure:

  0015-mungekey-cmdline.t 31 - mungekey without --force and with
    existing keyfile

Upon investigation, the contents of ${keyfile} were "-n xyzzy-31315"
(where the integer was the expansion of "$$" for the current pid).
These contents were generated with "echo -n", and macOS does not
appear to support the "-n" option for "echo".

The Autoconf Portable Shell Programming guide notes that it is
not possible to use "echo" portably unless both options and escape
sequences are omitted.

Remove the "-n" option from all instances of "echo".  None of these
instances require the omission of the trailing newline.  Since this
is the only "echo" option used, this change removes all options.
No "echo" backslash escape sequences are used.

Reference:
- https://www.gnu.org/savannah-checkouts/gnu/autoconf/manual/autoconf-2.72/html_node/Limitations-of-Builtins.html

Tested:
- AlmaLinux 9.5, 8.10
- Arch Linux
- CentOS Linux Stream 9, 8.5.2111, 7.9.2009, 6.10
- Debian sid, 12.10, 11.11, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 41, 40
- FreeBSD 14.2, 13.5
- NetBSD 10.1, 9.4
- OpenBSD 7.6, 7.5
- openSUSE 15.6, 15.5
- Ubuntu 24.10, 24.04.2, 22.04.5, 20.04.6, 18.04.6, 16.04.7, 14.04.6

Issue #141