Sharness: Add bail_out()
If mungekey or munged is unable to start due to a misconfiguration
(such as directory ownership or permissions), then many subsequent
checks in the test script will also fail.  (See Debian bugs #1094097
and #1102947 for examples of test-suite logs cluttered with failures
due to a directory ownership problem.)  As such, it is desirable to
bail out immediately.  The TAP specification includes the "Bail out!"
directive for just this reason.  Unfortunately, Sharness does not
implement it.

Add bail_out() (in "tests/sharness.d/00-bail_out.sh").  This is based
on git's BAIL_OUT() from "t/test-lib.sh".  It outputs to stdout the
magic words "Bail out!" followed by a reason, and then immediately
exits the test script (but continues on with any remaining test
scripts when using automake's "tap-driver.sh").

Additionally, update the test scripts to immediately bail out if the
initial mungekey or munged fail to run.

Example:

  $ mkdir -p /tmp/private
  $ chmod 700 /tmp/private
  $ TMPDIR=/tmp/private tests/0010-basic.t
  ok 1 - setup
  ok 2 - create key
  ok 3 - check keyfile creation
  Bail out! Failed to start munged

  Append the "-v" option to see the failing error message:

  $ TMPDIR=/tmp/private tests/0010-basic.t -v
  ...
  ok 3 - check keyfile creation

  expecting success:
      munged_start t-bail-out-on-error

  munged: Error: Socket is inaccessible: execute permissions for all required on "/tmp/private" (st_mode=040700)
  Bail out! Failed to start munged

Reference:
- https://testanything.org/tap-specification.html
- https://www.gnu.org/savannah-checkouts/gnu/automake/manual/html_node/Incompatibility-with-other-TAP-parsers-and-drivers.html
- https://github.com/git/git/blob/f65182a99e545d2f2bc22e6c1c2da192133b16a3/t/test-lib.sh#L672
- https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1094097
- https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1102947

Tested:
- AlmaLinux 9.5, 8.10
- Arch Linux
- CentOS Linux Stream 9, 8.5.2111, 7.9.2009, 6.10
- Debian sid, 12.10, 11.11, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 42, 41
- FreeBSD 14.2, 13.5
- NetBSD 10.1, 9.4
- OpenBSD 7.6, 7.5
- openSUSE 15.6, 15.5
- Ubuntu 25.04, 24.04.2, 22.04.5, 20.04.6, 18.04.6, 16.04.7, 14.04.6