Sharness: Update 0015-mungekey-cmdline.t to use MUNGE_KEYFILE
Replace the use of the local variable "keyfile" (which is set
relative to the sharness trash directory) with MUNGE_KEYFILE (defined
in munged_setup).  This is consistent with the other test scripts
and allows for MUNGE_KEYDIR or MUNGE_ROOT to be redefined in the
environment when running the test suite.

Tested:
- AlmaLinux 9.5, 8.10
- Arch Linux
- CentOS Linux Stream 9, 8.5.2111, 7.9.2009, 6.10
- Debian sid, 12.10, 11.11, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 42, 41
- FreeBSD 14.2, 13.5
- NetBSD 10.1, 9.4
- OpenBSD 7.6, 7.5
- openSUSE 15.6, 15.5
- Ubuntu 25.04, 24.04.2, 22.04.5, 20.04.6, 18.04.6, 16.04.7, 14.04.6