Sharness: Fix 1000-chaos-rpm.t merged-stream confusion
After commit fc4a925, "tests/1000-chaos-rpm.t" fails during test #5
(build rpm) on some RH-based systems when processed by automake's
"tap-driver.sh".  When the configure script is run during this step,
the following output is observed on some RH-based systems:

> ...
> checking for strip... strip
> checking command to parse /bin/nm -B output from gcc object... ./configure: line 7548:  -e 's/^T .* \(.*\)$/extern int \1();/p' -e 's/^[ABCDGIRSTW][ABCDGIRSTW]* .* \(.*\)$/extern char \1;/p': No such file or directory
> ok
> PASS: 1000-chaos-rpm.t 5
> checking for sysroot... no
> ...
> ok 5 - build rpm
> ERROR: 1000-chaos-rpm.t 5 - build rpm # OUT-OF-ORDER (expecting 6)

RH-based systems not exhibiting this error show the following output:

> ...
> checking for strip... strip
> checking command to parse /bin/nm -B output from gcc object... ok
> checking for sysroot... no
> ...
> ok 5 - build rpm

The TAP specification states:

  A harness must only read TAP output from standard output and not
  from standard error.  Lines written to standard output matching
  /^(not )?ok\b/ must be interpreted as test lines.

The autoconf "checking command to parse /bin/nm -B output from gcc
object" check seems to be encountering an error when constructing a
sed command.  The resulting error message causes the "ok" result to
be written to the start of the next line, and that "/^ok\b/" is being
misinterpreted by "tap-driver.sh" as a test result now that the test
script's stderr is being merged into stdout.  That misconstrued "ok"
causes the "OUT-OF-ORDER (expecting 6)" error shown above.

To address this, the rpmbuild "--pipe" option is used to pipe its
output to a sed command in order to rewrite the offending "ok" (e.g.,
"s/^ok\b/ ok/").  And that does seem to fix the failure, but not for
the expected reason.  With "--pipe", the rpmbuild-initiated configure
script in check #5 (build rpm) produces the following output on all
RH-based test systems:

> checking command to parse /bin/nm -B output from gcc object... ok

As such, the sed substitution never triggers.  In fact, changing
the substitution to "s/foo/bar/" also fixes the failure, but alas,
I'm unsure as to why.  I thought maybe rpmbuild could be causing
some mischief.  But Fedora 41 and 42 include the same version of
rpm (rpm-4.20.1-1), and F41 exhibits this failure but F42 does not.
Curiouser and curiouser.

Reference:
- https://testanything.org/tap-specification.html

Tested:
- AlmaLinux 9.5, 8.10
- CentOS Linux Stream 9, 8.5.2111
- Fedora 42, 41, 40