GitHub: Add build workflow
Create a GitHub Actions workflow to build and test on ubuntu-latest
and macos-latest with openssl and libgcrypt using gcc and clang.

The "fail-fast" strategy is disabled to prevent in-progress and
queued jobs in the matrix from being canceled if any matrix job fails.
"continue-on-error" is disabled in order for the workflow run to fail
if any matrix job fails.

The macOS build using libgcrypt is excluded due to the following
error during compilation:

> /opt/homebrew/Cellar/libgcrypt/1.11.0/include/gcrypt.h:31:10: fatal error: 'gpg-error.h' file not found

On macOS, "/tmp" is replaced with "/private/tmp" during "make check"
due to the following two sharness test failures without it:

  0101-munged-security-socket.t 18 - socket dir inaccessible by all failure
  0101-munged-security-socket.t 19 - socket dir inaccessible by all override

These checks grep the munged log for the fatal error message, and
their regex includes the expected pathname.  However, the pathname in
that log message is canonicalized with realpath().  When "make check"
was run with TMPDIR="/tmp", the socketdir pathname was rewritten
to start with "/private/tmp".  Presumably, "/tmp" is a symlink to
"/private/tmp" on macOS:

> munged: Error: Socket is inaccessible: execute permissions for all required on "/private/tmp/munge-32170/socketdir-32170" (st_mode=040700)

Setting TMPDIR="/private/tmp" avoids the symlink so the grep regex
pattern matches the log message thereby resolving this discrepancy.

Reference:
- https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/running-variations-of-jobs-in-a-workflow
- https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#jobsjob_idstrategyfail-fast
- https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#jobsjob_idcontinue-on-error
- https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/evaluate-expressions-in-workflows-and-actions#failure

Tested:
- macos-latest (14.7.5, Build 23H527, Darwin 23.6.0)
- ubuntu-latest (24.04.2)

Closes #141
Closes #142