Sharness: Fix grep regex for canonicalized paths
Commit cdae0a5 uncovered 2 failures in the test suite when it was run
with TMPDIR containing a symlink on macOS.  These failures occurred
when the pathname in the log message did not match the pathname in
the grep regex because the log message pathname had been canonicalized
by realpath().  This failure can also occur if the pathname contains
multiple sequential '/' characters.  When TEST_LONG=t is set, 8 tests
can fail in this manner:

  $ make check TEST_LONG=t TMPDIR=/tmp/
    (Note the trailing slash on TMPDIR's value)

  0101-munged-security-socket.t 9 - socket dir owned by other failure
  0101-munged-security-socket.t 10 - socket dir owned by other override
  0101-munged-security-socket.t 20 - socket dir inaccessible by all failure
  0101-munged-security-socket.t 21 - socket dir inaccessible by all override
  0102-munged-security-keyfile.t 26 - keyfile dir owned by other failure
  0102-munged-security-keyfile.t 27 - keyfile dir owned by other override
  0103-munged-security-logfile.t 23 - logfile dir owned by other failure
  0103-munged-security-logfile.t 24 - logfile dir owned by other override

However, the following 3 tests also have a regex pattern containing
a pathname but do not generate a failure:

  0100-munged-lock.t 9 - start munged with in-use socket
  0100-munged-lock.t 13 - start munged with 0600 bogus lockfile
  0100-munged-lock.t 14 - start munged with 0222 bogus lockfile

path_canonicalize() calls realpath(), and is called by
path_is_accessible() and path_is_secure().  path_is_secure() is
called for the keydir, logdir, piddir, sockdir, and PRNG seedpath.
But lock_create() does not call path_is_secure() for the socket
lockfile path (found in these last 3 tests which do not show failures)
because the socket lockfile must reside in the sockdir.

Update the grep regex patterns to omit pathnames in order to avoid
canonicalized path mismatches.

Tested:
- AlmaLinux 9.5, 8.10
- Arch Linux
- CentOS Linux Stream 9, 8.5.2111, 7.9.2009, 6.10
- Debian sid, 12.10, 11.11, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 42, 41
- FreeBSD 14.2, 13.5
- NetBSD 10.1, 9.4
- OpenBSD 7.7, 7.6
- openSUSE 15.6, 15.5
- Ubuntu 25.04, 24.04.2, 22.04.5, 20.04.6, 18.04.6, 16.04.7, 14.04.6