Cleanup: enc_armor: Replace strcpy with memcpy
In enc_armor(), replace both instances of strcpy() with memcpy().
This is a purely cosmetic change since the size of the malloc'd
destination buffer is sized to accommodate these strings.

The terminating null character is not copied in the first instance
because it would be immediately overwritten.  But it is copied in
the second instance in order to null-terminate the string at the end
of processing.