Cleanup: log_open_file: Replace strcpy with memccpy
In log_open_file(), replace strcpy() with memccpy() for copying
the identity string into the log_ctx.id array.  This changes the
behavior from silently ignoring the identity string if it exceeds
the array size to instead truncating it to fit the character array
(with guaranteed null-termination, of course).

Additionally, replace memset()ing the entire id array with just
writing a single null character at the start when identity is NULL.