Cleanup: _m_msg_client_connect: Replace strncpy with strnlen + memcpy
In _m_msg_client_connect(), replace strncpy() with strnlen() at the
top of the function followed later by a memcpy() that is guaranteed
to fit within the sockaddr_un addr.sun_path array.  This allows an
array overflow error to be returned immediately instead of later
attempting to copy a string that won't fit in the array.  strnlen()
protects against computing the length of an excessively long string.
And memcpy() avoids needlessly filling the addr.sun_path array with
trailing null bytes like strncpy() would.