Cleanup: sock_create: Replace strlcpy with strnlen + memcpy
In sock_create(), replace strlcpy() with strnlen() at the top of the
function followed later by a memcpy() that is guaranteed to fit within
the sockaddr_un addr.sun_path array.  This allows an array overflow
error to be returned immediately instead of later attempting to copy
a string that won't fit in the array.  strnlen() protects against
computing the length of an excessively long string unlike strlcpy()
which must return its strlen().