Cleanup: enc_armor: Remove unused int var
In enc_armor(), the integer variable "n" is first used to compute
the length of a packed credential before it is base64-encoded.  It is
then re-used to sum the number of bytes used during base64-encoding.
But this sum is never used once it has been computed, so remove it
since it is not needed.

Additionally, a second temporary integer variable "n2" is used during
base64-encoding.  Since "n" is no longer being used to calculate the
sum, replace "n2" with "n".