Cleanup: path.c: Replace strlcpy with strnlen + memcpy
In path_canonicalize() and path_dirname(), replace strlcpy() with
strnlen() + memcpy().  strnlen() protects against computing the length
of an excessively long src string -- unlike strlcpy() which must return
its strlen().  And if the string length exceeds the dst buffer size
(dstlen), immediately return an error instead of attempting to copy
a string that won't fit in the buffer -- unlike strlcpy().

Also for path_canonicalize(), return -1 (instead of strlen(src)) if
the src string is too long to fit in the dst buffer.  Update callers
path_is_accessible() and path_is_secure() to account for the changed
return value.