Spec: Remove libmunge.la on RHEL 10+
On AlmaLinux 10.0, rpmbuild fails with:

  File not found: .../usr/lib64/libmunge.la

When building rpms, Fedora 36+ removes libtool archives (".la" files)
from buildroot (see commit 787d1a9).  After installing files in
%install, the %__os_install_post scriptlet removes all ".la" files
using %__brp_remove_la_files.  RHEL 10 follows this behavior.

Update the specfile to exclude "libmunge.la" from the devel %files
on RHEL 10 or later (in addition to Fedora 36 or later).

Also fix the existing %if conditional expression for Fedora.  That
conditional ("%if 0%{?fedora} < 36") would be true on any non-Fedora
distro since it would evaluate to "%if 0 < 36".  The correct way to
perform this check is to add a guard to ensure that macro is defined
on the current platform:

  %if 0%{?fedora} && 0%{?fedora} < 36

A similar expression is used to check if %{?rhel} < 10.

Tested:
- AlmaLinux 10.0 (x86-64-v2), 9.6, 8.10
- CentOS Linux Stream 9, 8.5.2111, 7.9.2009
- Fedora 42, 41