tests: Add Valgrind suppression for getaddrinfo NSS leak
Add suppression for a memory leak in glibc's NSS (Name Service
Switch) module loading system that occurs during getaddrinfo()
hostname resolution.

The leak appears in the dynamic linker when NSS modules are loaded
for the first time during hostname lookups.  This is a one-time
initialization leak of 40 bytes in glibc's internal NSS infrastructure.

Observed on:
- Arch Linux (glibc 2.42+r17+gd7274d718e6f-1)
- Fedora 41 (glibc-2.40-27.fc41.x86_64)
- Fedora 42 (glibc-2.41-10.fc42.x86_64)