tests: Use unique stderr files to preserve error messages
In "0095-credential-payload.t", a test was failing for unexpected
reasons, but the actual error message was hidden by the subsequent
grep command that was checking for the expected error in that test.
Since all tests wrote stderr to the same "err.$$" file, each test
overwrote the previous stderr output, making it impossible to see
the actual error message for the failing test after script completion.

For each test, use a unique stderr file (errNN.$$) to preserve error
messages.  Update the "cleanup" routine to dump all stderr files
(in verbose mode) to help diagnose test failures.

Tested:
- AlmaLinux 10.0 (x86-64-v2), 9.6, 8.10
- Arch Linux
- CentOS Stream 9
- Debian sid, 13.1, 12.12, 11.11, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5, 4
- Fedora 42, 41
- FreeBSD 14.3, 13.5
- macOS 15.6
- NetBSD 10.1, 9.4
- OpenBSD 7.7, 7.6
- openSUSE Leap 15.6, 15.5
- Ubuntu 25.04, 24.04.3, 22.04.5, 20.04.6, 18.04.6, 16.04.7