spec: Update nologin path following /usr merge
Use "/usr/sbin/nologin" as the shell for the munge user account.

As part of the /usr merge implemented in modern distributions, /sbin
is now a symlink to /usr/sbin.  While both paths currently work due
to this symlink, "/usr/sbin/nologin" is the canonical location in
the unified filesystem hierarchy.

This aligns with:
- The /usr merge completed in Fedora 17 and adopted by other
  distributions (AlmaLinux, CentOS Stream)
- The systemd-sysusers default [see sysusers.d(5)]

Reference:
- https://systemd.io/THE_CASE_FOR_THE_USR_MERGE/
- https://fedoraproject.org/wiki/Features/UsrMove