build: Add systemd-sysusers support
On Fedora 43, an rpm built from the source tarball failed to install
with failed dependency errors for group(munge) and user(munge).

RPM 6.0 in Fedora 43 automatically generates virtual dependencies on
user(USERNAME) and group(GROUPNAME) when files in %files are owned by
specific users/groups.  To resolve these dependencies, packages must
now provide sysusers.d config files which are automatically parsed
to generate virtual provides.

The Fedora packaging guidelines recommend adding this config file as
"Source3" and using the %sysusers_create_compat macro to process
it during %pre.  However, this approach is designed for downstream
package maintainers and doesn't work for upstream tarballs used with
"rpmbuild -tb", since package sources are not available in the %pre
section which runs before files are unpacked to the system.

Instead, provide inline content in "munge.spec.in" and pass it to
systemd-sysusers (available since systemd 215) as a here-document.
This approach bypasses the compatibility macros (which are not present
in AlmaLinux 8) and is compatible with all currently supported
RHEL-based systems (running systemd 239 or later).  Specify the
"--replace=PATH" option since files belonging to the package are
not yet available on disk.  Replace "Requires(pre): shadow-utils"
with /usr/bin/systemd-sysusers.  Remove "exit 0" from %pre since
systemd-sysusers is idempotent and should propagate failures.

The sysusers configuration explicitly specifies "/usr/sbin/nologin"
as the login shell (6th field), matching the useradd invocation it
replaces.  This field is supported since systemd 238; all currently
supported systems run systemd 239 or later.

Add "--with-systemdsysusersdir" to the configure script to explicitly
specify the installation path for the sysusers.d config file.
Update the makefile to install this config file.

Reference:
- https://yselkowitz.github.io/blog/2025/06/26/f43-mass-rebuild.html
- https://fedoraproject.org/wiki/Changes/RPMSuportForSystemdSysusers
- https://docs.fedoraproject.org/en-US/packaging-guidelines/UsersAndGroups/
- systemd-sysusers(8)
- sysusers.d(5)

Tested:
- AlmaLinux 10.0 (x86-64-v2), 9.6, 8.10
- CentOS Stream 9
- Fedora 43, 42